+++ lib/dependency-manager-adapters/bower.js
@@ -4,16 +4,15 @@ var CoreObject    = require('core-object');
 var fs            = require('fs-extra');
 var RSVP          = require('rsvp');
 var path          = require('path');
-var findEmberPath = require('./find-ember-path');
 var extend        = require('extend');
 var rimraf        = RSVP.denodeify(require('rimraf'));
 var resolve       = RSVP.denodeify(require('resolve'));
-
+var findEmberPath = require('../utils/find-ember-path');
 
 module.exports = CoreObject.extend({
   init: function() {
     this._super.apply(this, arguments);
-    this.run = this.run || require('./run');
+    this.run = this.run || require('../utils/run');
   },
   bowerJSONFileName: 'bower.json',
   bowerJSONBackupFileName: 'bower.json.ember-try',
