+++ lib/dependency-manager-adapters/npm.js
@@ -10,7 +10,7 @@ var rimraf        = RSVP.denodeify(require('rimraf'));
 module.exports = CoreObject.extend({
   init: function() {
     this._super.apply(this, arguments);
-    this.run = this.run || require('./run');
+    this.run = this.run || require('../utils/run');
   },
   configKey: 'npm',
   packageJSON: 'package.json',
