+++ lib/utils/dependency-manager-adapter-factory.js
@@ -1,7 +1,7 @@
 'use strict';
 
-var BowerAdapter = require('./bower-adapter');
-var NpmAdapter   = require('./npm-adapter');
+var BowerAdapter = require('../dependency-manager-adapters/bower');
+var NpmAdapter   = require('../dependency-manager-adapters/npm');
 
 module.exports = {
   generateFromConfig: function(config, root) {
