+++ test/tasks/reset-test.js
@@ -7,8 +7,8 @@ var fs            = require('fs-extra');
 var fixtureBower  = require('../fixtures/bower.json');
 var writeJSONFile = require('../helpers/write-json-file');
 
-var remove = RSVP.denodeify(fs.remove);
-var root = process.cwd();
+var remove  = RSVP.denodeify(fs.remove);
+var root    = process.cwd();
 var tmproot = path.join(root, 'tmp');
 var tmpdir;
