+++ test/utils/scenario-manager-test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var expect          = require('chai').expect;
-var ScenarioManager = require('../lib/utils/scenario-manager');
+var ScenarioManager = require('../../lib/utils/scenario-manager');
 var CoreObject      = require('core-object');
 var RSVP            = require('rsvp');
